requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cffi==1.16.0
chardet==5.2.0
cssselect2==0.7.0
diff-match-patch==20230430
Django==5.0.3
django-import-export==4.0.7
django-phonenumber-field==7.3.0
fonttools==4.53.0
html5lib==1.1
pillow==10.3.0
pip==24.0
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
reportlab==4.2.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.3.0
tzdata==2024.1
weasyprint==62.2
webencodings==0.5.1
wheel==0.43.0
zopfli==0.2.3
